\name{vclTest-package}
\alias{vclTest-package}
\alias{vclTest}
\docType{package}
\title{
\packageTitle{vclTest}
}
\description{
\packageDescription{vclTest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{vclTest}
\packageIndices{vclTest}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{vclTest}

Maintainer: \packageMaintainer{vclTest}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
